requirements.txt
django
djangorestframework
djangorestframework-simplejwt
openai
stripe
